pyproject.toml
[tool.poetry]
name = "python-ha-project"
version = "0.1.0"
description = ""
authors = ["chawanghyeon"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.0"
onnxruntime = "^1.17.3"
pillow = "^10.3.0"
locust = "^2.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
